Optimized PutImageData and GetImageData functions (not fully tested)
TODO: test with alpha between 1 and 254.  Also test all combinations of
scalar / SSE2 / SSSE3.

TODO: portability.  I think this still depends on a GNU C endian macro.

TODO: massively tidy-up the code and comments.  Still full of experiments
and alternate ways to get better asm.

----

RGBA -> ARGB is just a rotate by 8 bits.  Native to big-endian is usually
just a byte-swap (or no-op), and we use htobe32 / be32toh for that.

The scalar version is probably about twice as fast, on Intel Haswell or
Skylake.

The SSE2 and SSSE3 versions are even faster.